Streaming decompress-gzip-upload, staged import
Download the entire DBR zip, but then setup a stream to decompress,
recompress with gzip, and upload to the S3 staging bucket. The stream
is generally bound by the speed of decompression, so in the time it
takes to simply unzip, we already have the gzipped DBR on S3.

Redshift can COPY FROM GZIP, so from that point, it’s easier. Create a
staging table with the same schema as line_items but lacking the
statement_date column. COPY FROM the data into it, then add the column
back with a default, and insert everything from there into line_items.