Use TRUNCATE instead of DELETE for month-to-date
With TRUNCATE, the deletion of all records in month_to_date is immedate (and cannot happen in a transaction). The downside is that without the transaction, month_to_date can be left empty if something breaks. The upside is that the subsequent VACUUM after importing should be faster, because there's no need to vacuum away the rows that were deleted.